refactor: added Bill Summary component
