+++ src/navigation/AppNavigation.tsx
@@ -15,8 +15,8 @@ import {horizontalScale} from '../assets/scaling';
 import Categories from '../screens/Categories';
 import Home from '../screens/Home';
 import OrderSuccess from '../screens/Orders/Order-Success';
-import SavedAddress from '../screens/Address/saved_address';
-import AddAddress from '../screens/Address/add_address';
+import SavedAddress from '../screens/Address/SavedAddress';
+import AddAddress from '../screens/Address/AddAddress';
 import Payment from '../screens/Payment';
 import Search from '../screens/Search';
 import {SvgXml} from 'react-native-svg';
